created models.py in backend/app/database
moved all database models there from main.py to make it cleaner
adjusted main.py with the import for it and added comments + precheck before entering a summoner to db
