+++ backend/app/main.py
@@ -1,64 +1,32 @@
+import os
 from fastapi import FastAPI
 from fastapi.middleware.cors import CORSMiddleware
-from sqlmodel import Field, Relationship, SQLModel, select
+from sqlmodel import SQLModel, select
 from sqlalchemy.ext.asyncio import create_async_engine, AsyncSession
 from typing import List, Optional
 from dotenv import load_dotenv
 # (Make sure riot_api.py is in backend/app/services/)
+# (make sure models.py is in backend/app/database/ )
+from app.database.models import Summoners
 from app.services.riot_api import get_puuid_by_riot_id
 
-import os
+
 
 load_dotenv()
 
 
-# --- 1. MODEL DEFINITIONS (The DB Schema) ---
-
-class Champions(SQLModel, table=True):
-    champion_id: int = Field(primary_key=True)
-    name: str
-    tags: str
-    participants: List["Participants"] = Relationship(back_populates="champion")
-
-class Summoners(SQLModel, table=True):
-    puuid: str = Field(primary_key=True)
-    game_name: str
-    tag_line: str
-    summoner_level: int
-    participations: List["Participants"] = Relationship(back_populates="summoner")
-
-class Matches(SQLModel, table=True):
-    match_id: str = Field(primary_key=True)
-    game_version: str
-    game_duration: int
-    queue_id: int
-    participants: List["Participants"] = Relationship(back_populates="match")
-
-class Participants(SQLModel, table=True):
-    internal_id: Optional[int] = Field(default=None, primary_key=True)
-    match_id: str = Field(foreign_key="matches.match_id")
-    puuid: str = Field(foreign_key="summoners.puuid")
-    champion_id: int = Field(foreign_key="champions.champion_id")
-    win: bool
-    kills: int
-    deaths: int
-    assists: int
-    individual_position: str
-
-    match: "Matches" = Relationship(back_populates="participants")
-    summoner: "Summoners" = Relationship(back_populates="participations")
-    champion: "Champions" = Relationship(back_populates="participants")
-
-# --- 2. DATABASE & APP SETUP ---
-#(Neo: added the database setup above and this heading to help separate of below code.)
+# DATABASE & APP SETUP
+#(Neo: Database  models are now in a separate file to keep main.py cleaner. See models.py for details and comments on the database structure.)
 
 app = FastAPI(title="Vantage Point Backend")
 
 # Get the URL from the docker-compose environment variable
+# poinmts to the db service not localhost hopfully, this should only work inside the container.
 DATABASE_URL = os.getenv("DATABASE_URL")
 engine = create_async_engine(DATABASE_URL)
 
 # CORS for frontend
+# 3000 = React default, 5173 = Vite default.
 app.add_middleware(
     CORSMiddleware,
     allow_origins=["http://localhost:3000", "http://localhost:5173"],
@@ -67,12 +35,19 @@ app.add_middleware(
     allow_headers=["*"],
 )
 
+
+# STARTUP
+
 @app.on_event("startup")
 async def on_startup():
-    # This automatically creates the tables in Postgres when the container starts
+    # Creates any tables that don't exist yet. Safe to run on every boot
+    # It won't touch tables that are already there.
     async with engine.begin() as conn:
         await conn.run_sync(SQLModel.metadata.create_all)
-    print("Database is synced and tables are ready!")
+    print("Tables are ready.")
+
+
+# ROUTES
 
 @app.get("/")
 async def root():
@@ -92,16 +67,21 @@ async def test_endpoint(data: dict):
 #Neo
 @app.post("/summoners/register")
 async def register_summoner(game_name: str, tag_line: str):
-    # 1. Get PUUID from Riot Service
+    # 1. Get PUUID from Riot Service; Gets name + tag 
     puuid = await get_puuid_by_riot_id(game_name, tag_line)
     if not puuid:
         return {"error": "Could not find player on Riot servers."}
 
-    # 2. Save to Database
+    # 2. Save to Database; should only do so if this player is not in the DB already
     async with AsyncSession(engine) as session:
         statement = select(Summoners).where(Summoners.puuid == puuid)
         result = await session.execute(statement)
         existing_summoner = result.scalar_one_or_none()
+
+        #adding this check just to be safe and security even if no exist is already below it
+        if existing_summoner:
+            return {"message": "Summoner already in database."}
+
 
         if not existing_summoner:
             new_summoner = Summoners(
@@ -114,4 +94,5 @@ async def register_summoner(game_name: str, tag_line: str):
             await session.commit()
             return {"message": f"Successfully registered {game_name}#{tag_line}", "puuid": puuid}
         
+        # should not be reached as the check i added earlier should catch this but just in case, 
         return {"message": "Summoner already in database."}
